Renamed method keys to tokens
