+++ src/index.ts
@@ -1,8 +1,13 @@
 export { VueSignalR, useSignalR } from './plugin';
 export { SignalRService } from './service';
-export { SignalRConfig } from './models/SignalRConfig';
-export {
-	SignalRServerMethod,
-	SignalRClientMethod
-} from './models/SignalRMethods';
+export { SignalRConfig } from './config';
+export { HubCommandToken, HubEventToken } from './tokens';
 export { SignalRSymbol } from './symbols';
+
+import { HubCommandToken, HubEventToken } from './tokens';
+
+/** @deprecated Use HubEventToken instead */
+export type SignalRClientMethod<T> = HubEventToken<T>;
+
+/** @deprecated Use HubCommandToken instead */
+export type SignalRServerMethod<T> = HubCommandToken<T>;
