--- src/models/SignalRMethods.ts
@@ -1,2 +1,0 @@
-export interface SignalRServerMethod<T> extends String {}
-export interface SignalRClientMethod<T> extends String {}
