+++ src/tokens.ts
@@ -1,0 +1,2 @@
+export interface HubCommandToken<T> extends String {}
+export interface HubEventToken<T> extends String {}
